+++ unix_integration/pam_kanidm/src/core.rs
@@ -102,82 +102,6 @@ pub trait PamHandler {
     fn prompt_for_mfacode(&self) -> PamResult<Option<String>>;
 }
 
-impl PamHandler for PamHandle {
-    fn account_id(&self) -> PamResult<String> {
-        self.get_user(None)
-    }
-
-    fn service_info(&self) -> PamResult<PamServiceInfo> {
-        self.get_pam_info()
-    }
-
-    fn consume_authtok(&self) -> PamResult<Option<String>> {
-        todo!();
-        /*
-        let mut consume_authtok = None;
-        std::mem::swap(&mut authtok, &mut consume_authtok);
-        match pamh.get_authtok() {
-            Ok(Some(v)) => Some(v),
-            Ok(None) => {
-                if opts.use_first_pass {
-                    debug!("Don't have an authtok, returning PAM_AUTH_ERR");
-                    return PamResultCode::PAM_AUTH_ERR;
-                }
-                None
-            }
-            Err(e) => {
-                error!(err = ?e, "get_authtok");
-                return e;
-            }
-        };
-        */
-    }
-
-    fn message(&self, prompt: &str) -> PamResult<()> {
-        todo!();
-    }
-
-    fn prompt_for_password(&self) -> PamResult<Option<String>> {
-        /*
-        let conv = match pamh.get_item::<PamConv>() {
-            Ok(conv) => conv,
-            Err(err) => {
-                error!(?err, "pam_conv");
-                return err;
-            }
-        };
-        conv.send(PAM_PROMPT_ECHO_OFF, "Password: ")
-        */
-        todo!();
-    }
-
-    fn prompt_for_mfacode(&self) -> PamResult<Option<String>> {
-        /*
-        conv.send(PAM_PROMPT_ECHO_OFF, "Code: ")
-        */
-        todo!();
-    }
-
-    fn prompt_for_pin(&self) -> PamResult<Option<String>> {
-        /*
-        conv.send(PAM_PROMPT_ECHO_OFF, "PIN: ")
-        */
-        todo!();
-    }
-
-    fn message_device_grant(&self, data: &DeviceAuthorizationResponse) -> PamResult<()> {
-        /*
-        let msg = match &data.message {
-            Some(msg) => msg.clone(),
-            None => format!("Using a browser on another device, visit:\n{}\nAnd enter the code:\n{}",
-                            data.verification_uri, data.user_code)
-        };
-        conv.send(PAM_TEXT_INFO, &msg)
-        */
-        todo!();
-    }
-}
-
 pub fn sm_authenticate_connected<P: PamHandler>(
     pamh: &P,
     opts: &ModuleOptions,
